refactor grpc
