+++ src/CraneCtld/CMakeLists.txt
@@ -1,13 +1,11 @@
+
+add_subdirectory(RpcService)
 add_executable(cranectld
         CtldPublicDefs.h
-        CtldGrpcServer.h
-        CtldGrpcServer.cpp
         DbClient.h
         DbClient.cpp
         TaskScheduler.h
         TaskScheduler.cpp
-        CranedKeeper.h
-        CranedKeeper.cpp
         CranedMetaContainer.h
         CranedMetaContainer.cpp
         AccountManager.h
@@ -36,6 +34,8 @@ target_link_libraries(cranectld PRIVATE
 
         Utility_PublicHeader
         Utility_PluginClient
+
+        RpcService
 
         uvw
