+++ src/CraneCtld/RpcService/CtldGrpcServer.cpp
@@ -26,6 +26,7 @@
 #include "TaskScheduler.h"
 #include "crane/PluginClient.h"
 #include "protos/PublicDefs.pb.h"
+#include "CtldForInternalServer.h"
 
 namespace Ctld {
 
@@ -40,7 +41,7 @@ grpc::Status CraneCtldServiceImpl::SubmitBatchTask(
   auto task = std::make_unique<TaskInCtld>();
   task->SetFieldsByTaskToCtld(request->task());
 
-  auto result = m_ctld_server_->SubmitTaskToScheduler(std::move(task));
+  auto result = g_task_scheduler->SubmitTaskToScheduler(std::move(task));
   if (result.has_value()) {
     task_id_t id = result.value().get();
     if (id != 0) {
@@ -76,7 +77,7 @@ grpc::Status CraneCtldServiceImpl::SubmitBatchTasks(
     auto task = std::make_unique<TaskInCtld>();
     task->SetFieldsByTaskToCtld(task_to_ctld);
 
-    auto result = m_ctld_server_->SubmitTaskToScheduler(std::move(task));
+    auto result = g_task_scheduler->SubmitTaskToScheduler(std::move(task));
     results.emplace_back(std::move(result));
   }
 
@@ -1403,203 +1404,6 @@ grpc::Status CraneCtldServiceImpl::EnableAutoPowerControl(
   return grpc::Status::OK;
 }
 
-grpc::Status CraneCtldServiceImpl::CforedStream(
-    grpc::ServerContext *context,
-    grpc::ServerReaderWriter<crane::grpc::StreamCtldReply,
-                             crane::grpc::StreamCforedRequest> *stream) {
-  if (!g_runtime_status.srv_ready.load(std::memory_order_acquire))
-    return grpc::Status{grpc::StatusCode::UNAVAILABLE,
-                        "CraneCtld Server is not ready"};
-
-  using crane::grpc::InteractiveTaskType;
-  using crane::grpc::StreamCforedRequest;
-  using crane::grpc::StreamCtldReply;
-  using grpc::Status;
-
-  enum class StreamState {
-    kWaitRegReq = 0,
-    kWaitMsg,
-    kCleanData,
-  };
-
-  bool ok;
-
-  StreamCforedRequest cfored_request;
-
-  auto stream_writer = std::make_shared<CforedStreamWriter>(stream);
-  std::weak_ptr<CforedStreamWriter> writer_weak_ptr(stream_writer);
-  std::string cfored_name;
-
-  CRANE_TRACE("CforedStream from {} created.", context->peer());
-
-  StreamState state = StreamState::kWaitRegReq;
-  while (true) {
-    switch (state) {
-    case StreamState::kWaitRegReq:
-      ok = stream->Read(&cfored_request);
-      if (ok) {
-        if (cfored_request.type() != StreamCforedRequest::CFORED_REGISTRATION) {
-          CRANE_ERROR("Expect type CFORED_REGISTRATION from peer {}.",
-                      context->peer());
-          return Status::CANCELLED;
-        }
-
-        cfored_name = cfored_request.payload_cfored_reg().cfored_name();
-        CRANE_INFO("Cfored {} registered.", cfored_name);
-
-        ok = stream_writer->WriteCforedRegistrationAck({});
-        if (ok) {
-          state = StreamState::kWaitMsg;
-        } else {
-          CRANE_ERROR(
-              "Failed to send msg to cfored {}. Connection is broken. "
-              "Exiting...",
-              cfored_name);
-          state = StreamState::kCleanData;
-        }
-
-      } else {
-        state = StreamState::kCleanData;
-      }
-
-      break;
-
-    case StreamState::kWaitMsg: {
-      ok = stream->Read(&cfored_request);
-      if (ok) {
-        switch (cfored_request.type()) {
-        case StreamCforedRequest::TASK_REQUEST: {
-          auto const &payload = cfored_request.payload_task_req();
-          auto task = std::make_unique<TaskInCtld>();
-          task->SetFieldsByTaskToCtld(payload.task());
-
-          auto &meta = std::get<InteractiveMetaInTask>(task->meta);
-
-          meta.cb_task_res_allocated =
-              [writer_weak_ptr](task_id_t task_id,
-                                std::string const &allocated_craned_regex,
-                                std::list<std::string> const &craned_ids) {
-                if (auto writer = writer_weak_ptr.lock(); writer)
-                  writer->WriteTaskResAllocReply(
-                      task_id,
-                      {std::make_pair(allocated_craned_regex, craned_ids)});
-              };
-
-          meta.cb_task_cancel = [writer_weak_ptr](task_id_t task_id) {
-            CRANE_TRACE("Sending TaskCancelRequest in task_cancel", task_id);
-            if (auto writer = writer_weak_ptr.lock(); writer)
-              writer->WriteTaskCancelRequest(task_id);
-          };
-
-          meta.cb_task_completed = [this, cfored_name, writer_weak_ptr](
-                                       task_id_t task_id,
-                                       bool send_completion_ack) {
-            CRANE_TRACE("The completion callback of task #{} has been called.",
-                        task_id);
-            if (auto writer = writer_weak_ptr.lock(); writer) {
-              if (send_completion_ack)
-                writer->WriteTaskCompletionAckReply(task_id);
-            } else {
-              CRANE_ERROR(
-                  "Stream writer of ia task #{} has been destroyed. "
-                  "TaskCompletionAckReply will not be sent.",
-                  task_id);
-            }
-
-            m_ctld_server_->m_mtx_.Lock();
-
-            // If cfored disconnected, the cfored_name should have be
-            // removed from the map and the task completion callback is
-            // generated from cleaning the remaining tasks by calling
-            // g_task_scheduler->TerminateTask(), we should ignore this
-            // callback since the task id has already been cleaned.
-            auto iter =
-                m_ctld_server_->m_cfored_running_tasks_.find(cfored_name);
-            if (iter != m_ctld_server_->m_cfored_running_tasks_.end())
-              iter->second.erase(task_id);
-            m_ctld_server_->m_mtx_.Unlock();
-          };
-
-          auto submit_result =
-              m_ctld_server_->SubmitTaskToScheduler(std::move(task));
-          std::expected<task_id_t, std::string> result;
-          if (submit_result.has_value()) {
-            result = std::expected<task_id_t, std::string>{
-                submit_result.value().get()};
-          } else {
-            result = std::unexpected(CraneErrStr(submit_result.error()));
-          }
-          ok = stream_writer->WriteTaskIdReply(payload.pid(), result);
-
-          if (!ok) {
-            CRANE_ERROR(
-                "Failed to send msg to cfored {}. Connection is broken. "
-                "Exiting...",
-                cfored_name);
-            state = StreamState::kCleanData;
-          } else {
-            if (result.has_value()) {
-              m_ctld_server_->m_mtx_.Lock();
-              m_ctld_server_->m_cfored_running_tasks_[cfored_name].emplace(
-                  result.value());
-              m_ctld_server_->m_mtx_.Unlock();
-            }
-          }
-        } break;
-
-        case StreamCforedRequest::TASK_COMPLETION_REQUEST: {
-          auto const &payload = cfored_request.payload_task_complete_req();
-          CRANE_TRACE("Recv TaskCompletionReq of Task #{}", payload.task_id());
-          if (g_task_scheduler->TerminatePendingOrRunningIaTask(
-                  payload.task_id()) != CraneErrCode::SUCCESS)
-            stream_writer->WriteTaskCompletionAckReply(payload.task_id());
-          else {
-            CRANE_TRACE(
-                "Termination of task #{} succeeded. "
-                "Leave TaskCompletionAck to TaskStatusChange.",
-                payload.task_id());
-          }
-        } break;
-
-        case StreamCforedRequest::CFORED_GRACEFUL_EXIT: {
-          stream_writer->WriteCforedGracefulExitAck();
-          stream_writer->Invalidate();
-          state = StreamState::kCleanData;
-        } break;
-
-        default:
-          CRANE_ERROR("Not expected cfored request type: {}",
-                      StreamCforedRequest_CforedRequestType_Name(
-                          cfored_request.type()));
-          return Status::CANCELLED;
-        }
-      } else {
-        state = StreamState::kCleanData;
-      }
-    } break;
-
-    case StreamState::kCleanData: {
-      CRANE_INFO("Cfored {} disconnected. Cleaning its data...", cfored_name);
-      stream_writer->Invalidate();
-      m_ctld_server_->m_mtx_.Lock();
-
-      auto const &running_task_set =
-          m_ctld_server_->m_cfored_running_tasks_[cfored_name];
-      std::vector<task_id_t> running_tasks(running_task_set.begin(),
-                                           running_task_set.end());
-      m_ctld_server_->m_cfored_running_tasks_.erase(cfored_name);
-      m_ctld_server_->m_mtx_.Unlock();
-
-      for (task_id_t task_id : running_tasks) {
-        g_task_scheduler->TerminateRunningTask(task_id);
-      }
-
-      return Status::OK;
-    }
-    }
-  }
-}
-
 CtldServer::CtldServer(const Config::CraneCtldListenConf &listen_conf) {
   m_service_impl_ = std::make_unique<CraneCtldServiceImpl>(this);
 
@@ -1647,6 +1451,8 @@ CtldServer::CtldServer(const Config::CraneCtldListenConf &listen_conf) {
     // condition will occur.
     g_craned_keeper->Shutdown();
 
+    g_ctld_for_internal_server->Shutdown();
+
     p_server->Shutdown(std::chrono::system_clock::now() +
                        std::chrono::seconds(1));
   });
@@ -1654,75 +1460,6 @@ CtldServer::CtldServer(const Config::CraneCtldListenConf &listen_conf) {
 
   signal(SIGINT, &CtldServer::signal_handler_func);
   signal(SIGTERM, &CtldServer::signal_handler_func);
-}
-
-CraneExpected<std::future<task_id_t>> CtldServer::SubmitTaskToScheduler(
-    std::unique_ptr<TaskInCtld> task) {
-  if (!task->password_entry->Valid()) {
-    CRANE_DEBUG("Uid {} not found on the controller node", task->uid);
-    return std::unexpected(CraneErrCode::ERR_INVALID_UID);
-  }
-  task->SetUsername(task->password_entry->Username());
-
-  {  // Limit the lifecycle of user_scoped_ptr
-    auto user_scoped_ptr =
-        g_account_manager->GetExistedUserInfo(task->Username());
-    if (!user_scoped_ptr) {
-      CRANE_DEBUG("User '{}' not found in the account database",
-                  task->Username());
-      return std::unexpected(CraneErrCode::ERR_INVALID_USER);
-    }
-
-    if (task->account.empty()) {
-      task->account = user_scoped_ptr->default_account;
-      task->MutableTaskToCtld()->set_account(user_scoped_ptr->default_account);
-    } else {
-      if (!user_scoped_ptr->account_to_attrs_map.contains(task->account)) {
-        CRANE_DEBUG(
-            "Account '{}' is not in the user account list when submitting the "
-            "task",
-            task->account);
-        return std::unexpected(CraneErrCode::ERR_USER_ACCOUNT_MISMATCH);
-      }
-    }
-  }
-
-  if (!g_account_manager->CheckUserPermissionToPartition(
-          task->Username(), task->account, task->partition_id)) {
-    CRANE_DEBUG(
-        "User '{}' doesn't have permission to use partition '{}' when using "
-        "account '{}'",
-        task->Username(), task->partition_id, task->account);
-    return std::unexpected(CraneErrCode::ERR_PARTITION_MISSING);
-  }
-
-  auto enable_res = g_account_manager->CheckIfUserOfAccountIsEnabled(
-      task->Username(), task->account);
-  if (!enable_res) {
-    return std::unexpected(enable_res.error());
-  }
-
-  auto result = g_meta_container->CheckIfAccountIsAllowedInPartition(
-      task->partition_id, task->account);
-  if (!result) return std::unexpected(result.error());
-
-  task->SetSubmitTime(absl::Now());
-
-  result = TaskScheduler::HandleUnsetOptionalInTaskToCtld(task.get());
-  if (result) result = TaskScheduler::AcquireTaskAttributes(task.get());
-  if (result) result = TaskScheduler::CheckTaskValidity(task.get());
-  if (result) {
-    auto res = g_account_meta_container->TryMallocQosResource(*task);
-    if (res != CraneErrCode::SUCCESS) {
-      CRANE_ERROR("The requested QoS resources have reached the user's limit.");
-      return std::unexpected(res);
-    }
-    std::future<task_id_t> future =
-        g_task_scheduler->SubmitTaskAsync(std::move(task));
-    return {std::move(future)};
-  }
-
-  return std::unexpected(result.error());
 }
 
 }  // namespace Ctld
