+++ src/CraneCtld/RpcService/CtldGrpcServer.h
@@ -31,152 +31,11 @@ using crane::grpc::Craned;
 using grpc::Channel;
 using grpc::Server;
 
-class CforedStreamWriter {
- private:
-  using Mutex = absl::Mutex;
-  using LockGuard = absl::MutexLock;
-
-  using StreamCtldReply = crane::grpc::StreamCtldReply;
-
- public:
-  explicit CforedStreamWriter(
-      grpc::ServerReaderWriter<crane::grpc::StreamCtldReply,
-                               crane::grpc::StreamCforedRequest> *stream)
-      : m_stream_(stream), m_valid_(true) {}
-
-  bool WriteTaskIdReply(pid_t calloc_pid,
-                        std::expected<task_id_t, std::string> res) {
-    LockGuard guard(&m_stream_mtx_);
-    if (!m_valid_) return false;
-
-    StreamCtldReply reply;
-    reply.set_type(StreamCtldReply::TASK_ID_REPLY);
-    auto *task_id_reply = reply.mutable_payload_task_id_reply();
-    if (res.has_value()) {
-      task_id_reply->set_ok(true);
-      task_id_reply->set_pid(calloc_pid);
-      task_id_reply->set_task_id(res.value());
-    } else {
-      task_id_reply->set_ok(false);
-      task_id_reply->set_pid(calloc_pid);
-      task_id_reply->set_failure_reason(std::move(res.error()));
-    }
-
-    return m_stream_->Write(reply);
-  }
-
-  bool WriteTaskResAllocReply(
-      task_id_t task_id,
-      std::expected<std::pair<std::string, std::list<CranedId>>, std::string>
-          res) {
-    LockGuard guard(&m_stream_mtx_);
-    if (!m_valid_) return false;
-
-    StreamCtldReply reply;
-    reply.set_type(StreamCtldReply::TASK_RES_ALLOC_REPLY);
-    auto *task_res_alloc_reply = reply.mutable_payload_task_res_alloc_reply();
-    task_res_alloc_reply->set_task_id(task_id);
-
-    if (res.has_value()) {
-      task_res_alloc_reply->set_ok(true);
-      task_res_alloc_reply->set_allocated_craned_regex(
-          std::move(res.value().first));
-      std::ranges::for_each(res.value().second,
-                            [&task_res_alloc_reply](const auto &craned_id) {
-                              task_res_alloc_reply->add_craned_ids(craned_id);
-                            });
-    } else {
-      task_res_alloc_reply->set_ok(false);
-      task_res_alloc_reply->set_failure_reason(std::move(res.error()));
-    }
-
-    return m_stream_->Write(reply);
-  }
-
-  bool WriteTaskCompletionAckReply(task_id_t task_id) {
-    LockGuard guard(&m_stream_mtx_);
-    if (!m_valid_) return false;
-    CRANE_TRACE("Sending TaskCompletionAckReply to cfored of task id {}",
-                task_id);
-    StreamCtldReply reply;
-    reply.set_type(StreamCtldReply::TASK_COMPLETION_ACK_REPLY);
-
-    auto *task_completion_ack = reply.mutable_payload_task_completion_ack();
-    task_completion_ack->set_task_id(task_id);
-
-    return m_stream_->Write(reply);
-  }
-
-  bool WriteTaskCancelRequest(task_id_t task_id) {
-    LockGuard guard(&m_stream_mtx_);
-    if (!m_valid_) return false;
-
-    StreamCtldReply reply;
-    reply.set_type(StreamCtldReply::TASK_CANCEL_REQUEST);
-
-    auto *task_cancel_req = reply.mutable_payload_task_cancel_request();
-    task_cancel_req->set_task_id(task_id);
-
-    return m_stream_->Write(reply);
-  }
-
-  bool WriteCforedRegistrationAck(const std::expected<void, std::string> &res) {
-    LockGuard guard(&m_stream_mtx_);
-    if (!m_valid_) return false;
-
-    StreamCtldReply reply;
-    reply.set_type(StreamCtldReply::CFORED_REGISTRATION_ACK);
-
-    auto *cfored_reg_ack = reply.mutable_payload_cfored_reg_ack();
-    if (res.has_value()) {
-      cfored_reg_ack->set_ok(true);
-    } else {
-      cfored_reg_ack->set_ok(false);
-      cfored_reg_ack->set_failure_reason(std::move(res.error()));
-    }
-
-    return m_stream_->Write(reply);
-  }
-
-  bool WriteCforedGracefulExitAck() {
-    LockGuard guard(&m_stream_mtx_);
-    if (!m_valid_) return false;
-
-    StreamCtldReply reply;
-    reply.set_type(StreamCtldReply::CFORED_GRACEFUL_EXIT_ACK);
-
-    auto *cfored_graceful_exit_ack = reply.mutable_payload_graceful_exit_ack();
-    cfored_graceful_exit_ack->set_ok(true);
-
-    return m_stream_->Write(reply);
-  }
-
-  void Invalidate() {
-    LockGuard guard(&m_stream_mtx_);
-    m_valid_ = false;
-  }
-
- private:
-  Mutex m_stream_mtx_;
-
-  bool m_valid_;
-
-  grpc::ServerReaderWriter<crane::grpc::StreamCtldReply,
-                           crane::grpc::StreamCforedRequest> *m_stream_
-      ABSL_GUARDED_BY(m_stream_mtx_);
-};
-
 class CtldServer;
 
 class CraneCtldServiceImpl final : public crane::grpc::CraneCtld::Service {
  public:
   explicit CraneCtldServiceImpl(CtldServer *server) : m_ctld_server_(server) {}
-
-  grpc::Status CforedStream(
-      grpc::ServerContext *context,
-      grpc::ServerReaderWriter<crane::grpc::StreamCtldReply,
-                               crane::grpc::StreamCforedRequest> *stream)
-      override;
 
   grpc::Status SubmitBatchTask(
       grpc::ServerContext *context,
@@ -188,21 +47,6 @@ class CraneCtldServiceImpl final : public crane::grpc::CraneCtld::Service {
       grpc::ServerContext *context,
       const crane::grpc::SubmitBatchTasksRequest *request,
       crane::grpc::SubmitBatchTasksReply *response) override;
-
-  grpc::Status TaskStatusChange(
-      grpc::ServerContext *context,
-      const crane::grpc::TaskStatusChangeRequest *request,
-      crane::grpc::TaskStatusChangeReply *response) override;
-
-  grpc::Status CranedTriggerReverseConn(
-      grpc::ServerContext *context,
-      const crane::grpc::CranedTriggerReverseConnRequest *request,
-      google::protobuf::Empty *response) override;
-
-  grpc::Status CranedRegister(
-      grpc::ServerContext *context,
-      const crane::grpc::CranedRegisterRequest *request,
-      crane::grpc::CranedRegisterReply *response) override;
 
   grpc::Status CancelTask(grpc::ServerContext *context,
                           const crane::grpc::CancelTaskRequest *request,
@@ -341,9 +185,6 @@ class CtldServer {
 
   inline void Wait() { m_server_->Wait(); }
 
-  CraneExpected<std::future<task_id_t>> SubmitTaskToScheduler(
-      std::unique_ptr<TaskInCtld> task);
-
  private:
   template <typename K, typename V,
             typename Hash = absl::container_internal::hash_default_hash<K>>
@@ -358,13 +199,9 @@ class CtldServer {
   std::unique_ptr<CraneCtldServiceImpl> m_service_impl_;
   std::unique_ptr<Server> m_server_;
 
-  Mutex m_mtx_;
-  HashMap<std::string /* cfored_name */, HashSet<task_id_t>>
-      m_cfored_running_tasks_ ABSL_GUARDED_BY(m_mtx_);
-
-  inline static std::mutex s_signal_cv_mtx_;
-  inline static std::condition_variable s_signal_cv_;
-  static void signal_handler_func(int) { s_signal_cv_.notify_one(); };
+  inline static std::mutex s_sigint_mtx;
+  inline static std::condition_variable s_sigint_cv;
+  static void signal_handler_func(int) { s_sigint_cv.notify_one(); };
 
   friend class CraneCtldServiceImpl;
 };
