+++ src/CraneCtld/TaskScheduler.cpp
@@ -1073,6 +1073,71 @@ CraneErrCode TaskScheduler::ChangeTaskPriority(task_id_t task_id,
   return CraneErrCode::SUCCESS;
 }
 
+CraneExpected<std::future<task_id_t>> TaskScheduler::SubmitTaskToScheduler(
+    std::unique_ptr<TaskInCtld> task) {
+  if (!task->password_entry->Valid()) {
+    CRANE_DEBUG("Uid {} not found on the controller node", task->uid);
+    return std::unexpected(CraneErrCode::ERR_INVALID_UID);
+  }
+  task->SetUsername(task->password_entry->Username());
+
+  {  // Limit the lifecycle of user_scoped_ptr
+    auto user_scoped_ptr =
+        g_account_manager->GetExistedUserInfo(task->Username());
+    if (!user_scoped_ptr) {
+      CRANE_DEBUG("User '{}' not found in the account database",
+                  task->Username());
+      return std::unexpected(CraneErrCode::ERR_INVALID_USER);
+    }
+
+    if (task->account.empty()) {
+      task->account = user_scoped_ptr->default_account;
+      task->MutableTaskToCtld()->set_account(user_scoped_ptr->default_account);
+    } else {
+      if (!user_scoped_ptr->account_to_attrs_map.contains(task->account)) {
+        CRANE_DEBUG(
+            "Account '{}' is not in the user account list when submitting the "
+            "task",
+            task->account);
+        return std::unexpected(CraneErrCode::ERR_USER_ACCOUNT_MISMATCH);
+      }
+    }
+  }
+
+  if (!g_account_manager->CheckUserPermissionToPartition(
+          task->Username(), task->account, task->partition_id)) {
+    CRANE_DEBUG(
+        "User '{}' doesn't have permission to use partition '{}' when using "
+        "account '{}'",
+        task->Username(), task->partition_id, task->account);
+    return std::unexpected(CraneErrCode::ERR_PARTITION_MISSING);
+  }
+
+  auto enable_res = g_account_manager->CheckIfUserOfAccountIsEnabled(
+      task->Username(), task->account);
+  if (!enable_res) {
+    return std::unexpected(enable_res.error());
+  }
+
+  auto result = g_meta_container->CheckIfAccountIsAllowedInPartition(
+      task->partition_id, task->account);
+  if (!result) return std::unexpected(result.error());
+
+  result = g_task_scheduler->AcquireTaskAttributes(task.get());
+
+  if (result) result = g_task_scheduler->CheckTaskValidity(task.get());
+
+  task->SetSubmitTime(absl::Now());
+
+  if (result) {
+    std::future<task_id_t> future =
+        g_task_scheduler->SubmitTaskAsync(std::move(task));
+    return {std::move(future)};
+  }
+
+  return std::unexpected(result.error());
+}
+
 CraneErrCode TaskScheduler::SetHoldForTaskInRamAndDb_(task_id_t task_id,
                                                       bool hold) {
   m_pending_task_map_mtx_.Lock();
