+++ src/CraneCtld/TaskScheduler.h
@@ -485,6 +485,9 @@ class TaskScheduler {
 
   CraneErrCode ChangeTaskPriority(task_id_t task_id, double priority);
 
+  CraneExpected <std::future<task_id_t>> SubmitTaskToScheduler(
+      std::unique_ptr<TaskInCtld> task);
+
   void TaskStatusChangeWithReasonAsync(uint32_t task_id,
                                        const CranedId& craned_index,
                                        crane::grpc::TaskStatus new_status,
