+++ src/Craned/Craned.cpp
@@ -196,8 +196,8 @@ void ParseConfig(int argc, char** argv) {
         std::exit(1);
       }
 
-      g_config.CraneCtldListenPort =
-          YamlValueOr(config["CraneCtldListenPort"], kCtldDefaultPort);
+      g_config.CraneCtldForInternalListenPort =
+          YamlValueOr(config["CraneCtldForInternalListenPort"], kCtldForInternalDefaultPort);
 
       if (config["Nodes"]) {
         for (auto it = config["Nodes"].begin(); it != config["Nodes"].end();
