+++ src/Craned/CranedPublicDefs.h
@@ -72,7 +72,7 @@ struct Config {
   bool CompressedRpc{};
 
   std::string ControlMachine;
-  std::string CraneCtldListenPort;
+  std::string CraneCtldForInternalListenPort;
   std::string CranedDebugLevel;
 
   std::filesystem::path CraneBaseDir;
