+++ src/Utilities/PublicHeader/include/crane/PublicHeader.h
@@ -44,7 +44,10 @@ using CraneExpected = std::expected<T, CraneErrCode>;
 template <typename T>
 using CraneExpectedRich = std::expected<T, CraneRichError>;
 
-inline const char* const kDefaultHost = "0.0.0.0";
+inline const char* kCtldDefaultPort = "10011";
+inline const char* kCranedDefaultPort = "10010";
+inline const char* kCforedDefaultPort = "10012";
+inline const char* kCtldForInternalDefaultPort = "10013";
 
 inline const char* const kCtldDefaultPort = "10011";
 inline const char* const kCranedDefaultPort = "10010";
